web/icon Move the icon module, take the first alt text's first letter as a default
