+++ ui/web/src/app/common/avatar/avatar.component.scss
@@ -15,14 +15,38 @@
  * limitations under the License.
  */
 
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
+@import 'src/app/common/colors';
 
-import {IconComponent} from './icon/icon.component';
+:host {
+  line-height: normal;
+}
 
-@NgModule({
-  declarations: [IconComponent],
-  exports: [IconComponent],
-  imports: [CommonModule],
-})
-export class DiscordModule {}
+div.icon-container {
+  width: inherit;
+  height: inherit;
+  box-sizing: border-box;
+  border-radius: 100px;
+  border: 1px solid material-color('grey', '300');
+
+  img {
+    max-width: 100%;
+    max-height: 100%;
+    border-radius: 100px;
+  }
+
+  div.letter-container {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+    border-radius: 100px;
+    background-color: material-color('purple', '700');
+
+    span {
+      text-align: center;
+      color: $white;
+      width: 2em;
+    }
+  }
+}
+
