--- ui/web/src/app/discord/icon/icon.component.html
@@ -1,1 +1,0 @@
-<img *ngIf="src" [src]="src" />
