+++ ui/web/src/app/header/header.component.html
@@ -5,7 +5,7 @@
   <ng-container *ngIf="profile">
     <button mat-button [matMenuTriggerFor]="guildMenu" class="guild-selection">
       <div class="button-wrapper">
-        <img [src]="selectedGuild?.icon_url" />
+        <app-avatar [src]="selectedGuild?.icon_url" [alt]="selectedGuild?.discord_name"></app-avatar>
         <span>{{ selectedGuild?.discord_name }}</span>
         <mat-icon class="subtle">keyboard_arrow_down</mat-icon>
       </div>
@@ -15,7 +15,7 @@
               (click)="onGuildSelected(relationship)"
               mat-menu-item>
         <div class="button-wrapper">
-          <img [src]="relationship.guild?.icon_url" />
+          <app-avatar [src]="relationship.guild?.icon_url" [alt]="relationship.guild?.discord_name"></app-avatar>
           <span>{{ relationship.guild?.discord_name }}</span>
         </div>
       </button>
@@ -25,7 +25,7 @@
   <ng-container *ngIf="profile; else login">
     <button mat-button [matMenuTriggerFor]="userMenu" class="user-profile">
       <div class="button-wrapper">
-        <img [src]="profile.icon_url" />
+        <app-avatar [src]="profile.icon_url" [alt]="profile.username"></app-avatar>
         <mat-icon class="subtle">keyboard_arrow_down</mat-icon>
       </div>
     </button>
